Add post edit
Create the edit view scaffolding of the post using make:view and build
Edit Post view for the authenticated user to edit the existing post.
Add the view to edit action.

Build update action to validate the user edits and update the database
with the edits.  Redirect to the list of posts.